Uses List<string> instead of ArrayLists
Fixed the load typo to Load

Changed the set function to have the subSection first. It's much more intuitive and looks more organized when setting multiple variables. It was visually awkward to seperate the key from the value.

Removed the ToString() calls.

Added using to the Save function to keep it consistant with the Load function.

Removed Close() from the Save and Load function, because using closes automatically.